Refactor skrafldb into facade with backend selection
Extract the NDB implementation from skrafldb.py into skrafldb_ndb.py
and create skrafldb_pg.py as a stub for the PostgreSQL backend. The
skrafldb.py module is now a thin facade that imports from either
backend based on the DATABASE_BACKEND environment variable.

Also simplify the PostgreSQL backend by removing the redundant
DatabaseSession wrapper class from connection.py - the backend
now directly manages its sessionmaker and engine.
